Advent of Code - Day 05
A short trip away for the weekend caused delays here. Quite a fun simple problem, although probably not the most optimised it still manages to complete in a couple seconds.

Part 1 - I initially thought about taking a DFS approach since we are guaranteed to reach an exit we could continually pop from top and when we reach an obstacle we simply shift our direction by 1 in our DIRECTIONS array

Part 2 - This is most likely not the most optimal approach but I was going to do cycle detection with floyds cycle detection but figured that using a visited set would just work the same with little changes from part 1. Only change is since we sometimes overlap a visited position we need to add in the current direction.