Advent of Code - Day 07
Quick and simple, thank you Advent of Code team for letting me enjoy the football this weekend.

Part 1 - We built a quick BFS since we are only constructing a binary tree. I considered optimisations like checking if we exceed the resultant value to exit early but the runtime is quick enough to not worry.

Part 2 - Thankfully our first solution enabled us to convert this binary tree to a tertiary tree with little performance drop.