Advent of Code - Day 08
Quite a simple one today thankfully.
Part 1 involved continuing the distance between a1 -> a2 and then equally a2 -> a1. This is stored in a set to calculate the unique set of anti-nodes.

Part 2 just extends the gradient to draw a line so we can use a while loop to get all the anti-nodes within the grid.