Advent of Code - Day 10
It's a dynamic christmas!

Part 1 we have been gifted a good old fashioned Depth First Search problem. This first part was quite simple the only issue is not double counting your nine's.

Part 2 the DFS paid off because it lends itself quite well to a dynamic programming. This problem became counting the number of distinct paths to an edge, quite a common leetcode problem, since we have quite a large grid it makes sense to use DP (memoisation) to save a lot of time since we will be computing over many paths. This could also be solved with a hash map but since we have a set grid size its quite easy to do it this way.