Advent of Code - Day 12
Back to BFS.

Part 1 was a simple BFS where I applied flood fill to prevent re-calculating a single garden plot. Perimeter counts are any edge which has a different value or is out of bounds.

Part 2 is a little more difficult, one approach I seen suggested on Reddit for counting sides is if we consider orthogonal pairs their are only a set number of possible combinations that create corners. Each corner becomes a side.