Advent of Code - Day 14
Simulate, Simulate and more simulations.

Part 1, we simulate each positions move across the x and y axis, since we wrap back values we can calculate all our moves then modulo the board size to simulate all the wraps back.

Part 2, this solution is a little less clean but I tried a few ideas and one paid off. The first idea was that for the christmas tree we might expect a straight line down the tree so I tried searching for that. The next idea was just to simulate the first few thousand and try stop when I see a shape I recognise. The third and final paid off which is the assumption they are creating the trees with no overlap and shifting from there which happened to pay off.

Checking Reddit though I am blown away by the Chinese Remainder Theorem.