Advent of Code - Day 17
Part 1 reminded me of writing my first ever CPU emulator for the CHIP-8 trying to make sure I got the register and opcode operations setup correctly, it was then just a case of running the opcodes in sequence.

Part 2, I initially tried just simply finding how much I have to increase A by the replicate the length of the original sequence. I found it had to be somewhere >10M. When I was doing this I started to look at how changing A actually impacts the output. One thing noted from the challenge was we are working on a 3 bit system so looking into that we can see that each time we increment each output only depends on the lowest 3 bits and we iterate over those 3 bits each time 0-8 since they are all the program uses to determine its output.